Use dynamic variables in CCS search
Add additional unit test

Improve search page by sending postcode in request

Rename ui components
