+++ application/search/ccs_comparison_search/service.py
