+++ application/ui/src/App.tsx
@@ -2,23 +2,23 @@ import React from "react";
 import { Route, Routes } from "react-router-dom";
 import { useAppSelector } from "./app/hooks";
 import {
-	ErrorPage,
-	ForgottenPasswordPage,
-	HomePage,
-	LoginPage,
-	MenuPage,
-	RegisterPage,
-	ResultsPage,
-	SearchPage,
+	CCSComparisonResults,
+	CCSComparisonSearch,
+	ForgottenPassword,
+	Home,
+	Login,
+	Menu,
+	PageNotFound,
+	Register,
 } from "./components";
 import {
 	BASE_PATH,
+	CCS_COMPARISON_RESULTS_PATH,
+	CCS_COMPARISON_SEARCH_PATH,
 	FORGOTTEN_PASSWORD_PATH,
 	LOGIN_PATH,
 	MENU_PATH,
 	REGISTER_PATH,
-	SEARCH_PATH,
-	SEARCH_RESULTS_PATH,
 } from "./constants/paths";
 import { selectLoggedIn } from "./slices/authSlice";
 
@@ -30,28 +30,31 @@ export class App extends React.Component {
 
 function ApplicationRouting(): JSX.Element {
 	const isLoggedIn = useAppSelector(selectLoggedIn);
+
 	const publicRoutes = (
 		<Routes>
-			<Route path={BASE_PATH} element={<HomePage />} />
-			<Route path={LOGIN_PATH} element={<LoginPage />} />
-			<Route path={REGISTER_PATH} element={<RegisterPage />} />
+			<Route path={BASE_PATH} element={<Home />} />
+			<Route path={LOGIN_PATH} element={<Login />} />
+			<Route path={REGISTER_PATH} element={<Register />} />
+			<Route path={FORGOTTEN_PASSWORD_PATH} element={<ForgottenPassword />} />
 			<Route
-				path={FORGOTTEN_PASSWORD_PATH}
-				element={<ForgottenPasswordPage />}
+				path={CCS_COMPARISON_SEARCH_PATH}
+				element={<CCSComparisonSearch />}
 			/>
-			<Route path={SEARCH_PATH} element={<SearchPage />} />
-			<Route path={SEARCH_RESULTS_PATH} element={<ResultsPage />} />
-			<Route path="*" element={<ErrorPage />} />
+			<Route
+				path={CCS_COMPARISON_RESULTS_PATH}
+				element={<CCSComparisonResults />}
+			/>
+			<Route path="*" element={<PageNotFound />} />
 		</Routes>
 	);
+
 	const privateRoutes = (
 		<Routes>
-			<Route path={BASE_PATH} element={<HomePage />} />
-			<Route path={LOGIN_PATH} element={<MenuPage />} />
-			<Route path={MENU_PATH} element={<MenuPage />} />
-			<Route path={SEARCH_PATH} element={<SearchPage />} />
-			<Route path={SEARCH_RESULTS_PATH} element={<ResultsPage />} />
-			<Route path="*" element={<ErrorPage />} />
+			<Route path={BASE_PATH} element={<Home />} />
+			<Route path={LOGIN_PATH} element={<Login />} />
+			<Route path={MENU_PATH} element={<Menu />} />
+			<Route path="*" element={<PageNotFound />} />
 		</Routes>
 	);
 	return isLoggedIn ? privateRoutes : publicRoutes;
