+++ application/ui/src/__test__/App.spec.tsx
@@ -1,7 +1,7 @@
 import { describe, expect, test } from "@jest/globals";
 import { screen } from "@testing-library/react";
 import { App } from "../App";
-import { expectedPageText } from "../components/homePage/__test__/homePage.spec";
+import { expectedPageText } from "../components/home/__test__/home.spec";
 import { renderWithProvidersAndRouter } from "./utils-for-tests";
 
 describe("App Public Routes Tests", () => {
