+++ application/ui/src/components/ccsComparisonResults/resultCard.css
