+++ application/ui/src/components/ccsComparisonSearch/__test__/dropDowns.spec.tsx
