+++ application/ui/src/components/ccsComparisonSearch/__test__/sharedSearchForm.spec.tsx
