+++ application/ui/src/components/ccsComparisonSearch/dropDowns.tsx
