+++ application/ui/src/components/ccsComparisonSearch/search.css
