+++ application/ui/src/components/ccsComparisonSearch/sharedSearchForm.tsx
