+++ application/ui/src/components/index.ts
@@ -1,19 +1,19 @@
-import ErrorPage from "./errorPage/errorPage";
-import ForgottenPasswordPage from "./forgottenPasswordPage/forgottenPasswordPage";
-import HomePage from "./homePage/homePage";
-import LoginPage from "./loginPage/loginPage";
-import MenuPage from "./menuPage/menuPage";
-import RegisterPage from "./registerPage/registerPage";
-import ResultsPage from "./resultsPage/resultsPage";
-import SearchPage from "./searchPage/searchPage";
+import CCSComparisonResults from "./ccsComparisonResults/ccsComparisonResults";
+import CCSComparisonSearch from "./ccsComparisonSearch/ccsComparisonSearch";
+import ForgottenPassword from "./forgottenPassword/forgottenPassword";
+import Home from "./home/home";
+import Login from "./login/login";
+import Menu from "./menu/menu";
+import PageNotFound from "./pageNotFound/pageNotFound";
+import Register from "./register/register";
 
 export {
-	RegisterPage,
-	HomePage,
-	LoginPage,
-	MenuPage,
-	ErrorPage,
-	ForgottenPasswordPage,
-	SearchPage,
-	ResultsPage,
+	CCSComparisonResults,
+	CCSComparisonSearch,
+	ForgottenPassword,
+	Home,
+	Login,
+	Menu,
+	PageNotFound,
+	Register,
 };
