+++ application/ui/src/constants/paths.ts
@@ -3,5 +3,5 @@ export const LOGIN_PATH = "/login";
 export const MENU_PATH = "/menu";
 export const REGISTER_PATH = "/register";
 export const FORGOTTEN_PASSWORD_PATH = "/forgotten-password";
-export const SEARCH_PATH = "/search";
-export const SEARCH_RESULTS_PATH = "/search-results";
+export const CCS_COMPARISON_SEARCH_PATH = "/ccs-comparison-search";
+export const CCS_COMPARISON_RESULTS_PATH = "/ccs-comparison-results";
